refactor: extract out verify release hook
